Python/01/requirements.txt
discord.py
mysql-connector-python
bs4
discord-Together
wavelink
discord_emoji